Add classic battle page wrapper
