+++ src/pages/randomJudoka.html
@@ -41,10 +41,20 @@
       <div class="card-section">
         <div id="card-container" data-testid="card-container" class="card-container"></div>
         <!-- PRD: Draw Card button and toggles per prdRandomJudoka.md & prdDrawRandomCard.md -->
-        <div id="draw-controls" class="draw-controls" aria-label="Draw controls" style="display: flex; flex-direction: column; align-items: center; margin-top: 24px;">
+        <div
+          id="draw-controls"
+          class="draw-controls"
+          aria-label="Draw controls"
+          style="display: flex; flex-direction: column; align-items: center; margin-top: 24px"
+        >
           <!-- Button and toggles are injected by randomJudokaPage.js -->
           <!-- Error message container for accessibility -->
-          <div id="draw-error-message" role="alert" aria-live="assertive" style="color: #b00020; font-size: 1.1rem; margin-top: 12px;"></div>
+          <div
+            id="draw-error-message"
+            role="alert"
+            aria-live="assertive"
+            style="color: #b00020; font-size: 1.1rem; margin-top: 12px"
+          ></div>
         </div>
       </div>
