+++ src/pages/settings.html
@@ -62,7 +62,13 @@
             <legend>General Settings</legend>
             <div class="settings-item">
               <label for="sound-toggle" class="switch">
-                <input type="checkbox" id="sound-toggle" name="sound" aria-label="Sound" tabindex="2" />
+                <input
+                  type="checkbox"
+                  id="sound-toggle"
+                  name="sound"
+                  aria-label="Sound"
+                  tabindex="2"
+                />
                 <div class="slider round"></div>
                 <span>Sound</span>
               </label>
@@ -105,7 +111,13 @@
             <a href="../pages/changeLog.html" id="changelog-link" tabindex="99">View Change Log</a>
           </div>
         </form>
-        <div id="settings-error-popup" class="settings-error-popup" role="alert" aria-live="assertive" style="display:none;"></div>
+        <div
+          id="settings-error-popup"
+          class="settings-error-popup"
+          role="alert"
+          aria-live="assertive"
+          style="display: none"
+        ></div>
       </main>
 
       <footer>
